refactor: moving client-app/ to experimental/
refactor: adding copyright headers

refactor: adding copyright header
