--- client-app/src/components/campaign/styles/campaignList.js
@@ -1,13 +1,0 @@
-import { css } from 'lit';
-
-const campaignStyles =  css`
-    .campaignList {
-      display: flex;
-      align-items: center;
-      justify-content: center;
-      flex-wrap: wrap;
-    }
-`;
-
-export default campaignStyles;
-
