--- client-app/src/stores/selectors/campaigns.js
@@ -1,5 +1,0 @@
-
-export const getCampaignItem = (store, campaignId) => {
-  const totalCampaigns = store.getState()?.campaigns?.campaigns; 
-  return totalCampaigns.filter(campaign => campaign.id === campaignId)?.[0];
-};
