+++ experimental/client-app/assets/cymbal-giving-logo.png
