+++ experimental/client-app/src/stores/base.js
@@ -1,0 +1,24 @@
+// Copyright 2022 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+import { configureStore } from '@reduxjs/toolkit';
+import campaignReducer from './slices/campaign.js';
+import campaignListReducer from './slices/campaignList.js';
+
+export default configureStore({
+  reducer: {
+    campaign: campaignReducer,
+    campaignList: campaignListReducer
+  }
+});
