+++ experimental/client-app/src/styles/shell.js
@@ -1,0 +1,46 @@
+// Copyright 2022 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+import { css } from 'lit';
+
+/* stylelint-disable font-family-no-missing-generic-family-keyword */
+const shellStyles =  css`
+  :host {
+    display: flex;
+    flex-direction: column;
+    font-family: 'Google Sans';
+  }
+
+  main {
+    margin: auto 20px;
+    flex-grow: 1;
+  }
+
+  app-header {
+    width: 100%;
+    height: 100%;
+  }
+
+  .title {
+    display: flex;
+    align-items: center;
+    font-size: 25px;
+    width: 100%;
+    padding: 30px;
+  }
+`;
+/* stylelint-enable font-family-no-missing-generic-family-keyword */
+
+export default shellStyles;
+
