+++ experimental/client-app/test/app-shell.test.js
@@ -1,0 +1,37 @@
+// Copyright 2022 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+// TODO: Part of open-wc scaffolding. Needs to be refactored to reflect actual app-shell
+
+import { html } from 'lit';
+import { fixture, expect } from '@open-wc/testing';
+
+import '../src/app-shell.js';
+
+describe('AppShell', () => {
+  let element;
+  beforeEach(async () => {
+    element = await fixture(html`<app-shell></app-shell>`);
+  });
+
+  it('renders a h1', () => {
+    const h1 = element.shadowRoot.querySelector('h1');
+    expect(h1).to.exist;
+    expect(h1.textContent).to.equal('My app');
+  });
+
+  it('passes the a11y audit', async () => {
+    await expect(element).shadowDom.to.be.accessible();
+  });
+});
